third_party/WebKit/LayoutTests/css3/device-adapt/opera/descriptor-values-005-expected.txt


FAIL CSS Test: @viewport zoom factor - allow zero factors assert_equals: expected 3 but got 1
